Verify the capture-core SQL against a real PostgreSQL cluster
Migration 0018 could never have applied: PostgreSQL has no
mod(double precision, integer), and the double-to-numeric cast is
assignment-only, so CREATE VIEW failed and v_dart_locations was never
created. Cast the angle to NUMERIC before the modulo and guard the shape
with a test over the whole migration chain.

Running the chain also let the deferred operator checklist steps happen:
the introspect run confirms the hand-written coordinate columns and
constraint match generated output, and the constraint, angle convention
and bust divergence checks all pass. Checklist rewritten to record what
is now verified and what still needs Neon.

+++ database/migrations/0018_dart_location_read_model.sql
@@ -38,7 +38,7 @@ SELECT es.id AS session_id,
     MOD(
         DEGREES(
             ATAN2(d.location_x, - d.location_y)
-        ) + 360,
+        )::NUMERIC + 360,
         360
     ) AS angle_degrees
 FROM darts d
